Remove old casts.cc
