+++ src/core/datatablemodule.cc
@@ -377,7 +377,6 @@ void py::DatatableModule::init_methods() {
   init_methods_str();
   init_methods_styles();
 
-  init_casts();
   init_fbinary();
   init_fnary();
   init_funary();
